932 refactor components and containers (#1341)
* refactor layout components

* refactor to fix lint errors

* refactor containers and components

* fix merge
