+++ packages/psss/src/components/Modal/index.js
@@ -4,8 +4,3 @@
  */
 
 export * from './AlertCreatedModal';
-export * from './CreateOutbreakModal';
-export * from './ExportModal';
-export * from './WeeklyReportsExportModal';
-export * from './OutbreaksExportModal';
-export * from './AlertsExportModal';
