+++ packages/psss/src/components/Table/index.js
@@ -3,10 +3,8 @@
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  */
 export * from './CellComponents';
-export * from './ConnectedTable';
 export * from './Table';
 export * from './TableHeader';
 export * from './TableBody';
 export * from './TableRow';
-export * from './VerifiableTableRow';
 export * from './dataAccessors';
