+++ packages/psss/src/components/index.js
@@ -13,6 +13,5 @@ export * from './DropdownMenu';
 export * from './SiteAddress';
 export * from './ActivityTab';
 export * from './FetchLoader';
-export * from './NotesTab';
 export * from './Modal';
 export * from './AffectedSitesTab';
