+++ packages/psss/src/containers/Modals/AlertsExportModal.js
@@ -2,13 +2,12 @@
  * Tupaia
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  */
-
 import React from 'react';
-import styled from 'styled-components';
 import PropTypes from 'prop-types';
+import styled from 'styled-components';
 import { Checkbox } from '@tupaia/ui-components';
 import { ExportModal } from './ExportModal';
-import { FlexStart } from '../Layout';
+import { FlexStart } from '../../components/Layout';
 
 // Todo replace with api data
 const syndromes = [
@@ -21,47 +20,36 @@ const syndromes = [
 
 const Checkboxes = styled(FlexStart)`
   margin-top: 1rem;
-  margin-bottom: 2rem;
   padding-top: 2rem;
   border-top: 1px solid ${props => props.theme.palette.grey['400']};
-  border-bottom: 1px solid ${props => props.theme.palette.grey['400']};
 
   > div {
     margin-right: 0.4rem;
   }
 `;
 
-export const WeeklyReportsExportModal = ({ isOpen, handleClose }) => (
+export const AlertsExportModal = ({ isOpen, handleClose }) => (
   <ExportModal
     isOpen={isOpen}
     handleClose={handleClose}
     renderCustomFields={register => (
-      <>
-        <Checkboxes>
-          {syndromes.map(syndrome => (
-            <Checkbox
-              key={syndrome.value}
-              label={syndrome.label}
-              name={syndrome.value}
-              color="primary"
-              defaultChecked
-              inputRef={register()}
-            />
-          ))}
-        </Checkboxes>
-        <Checkbox
-          name="includeAllSites"
-          label="Include all Sentinel sites"
-          color="primary"
-          defaultChecked
-          inputRef={register()}
-        />
-      </>
+      <Checkboxes>
+        {syndromes.map(syndrome => (
+          <Checkbox
+            key={syndrome.value}
+            label={syndrome.label}
+            name={syndrome.value}
+            color="primary"
+            defaultChecked
+            inputRef={register()}
+          />
+        ))}
+      </Checkboxes>
     )}
   />
 );
 
-WeeklyReportsExportModal.propTypes = {
+AlertsExportModal.propTypes = {
   isOpen: PropTypes.bool.isRequired,
   handleClose: PropTypes.func.isRequired,
 };
