+++ packages/psss/src/containers/index.js
@@ -6,4 +6,5 @@
 export * from './Tables';
 export * from './Forms';
 export * from './Panels';
+export * from './Modals';
 export * from './NavBar';
