+++ packages/psss/src/views/AlertsOutbreaksView.js
@@ -9,10 +9,9 @@ import { Archive } from '@material-ui/icons';
 import {
   Header,
   HeaderTitle,
-  OutbreaksExportModal,
-  AlertsExportModal,
   HeaderTitleWithSubHeading,
 } from '../components';
+import { AlertsExportModal, OutbreaksExportModal } from '../containers/Modals';
 import { AlertsRoutes } from '../routes/AlertsRoutes';
 import { countryFlagImage } from '../utils';
