+++ packages/psss/src/views/CountryReportsView.js
@@ -8,7 +8,8 @@ import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { useParams } from 'react-router-dom';
 import { TabsToolbar, CalendarToday } from '@tupaia/ui-components';
-import { Header, HeaderAvatarTitle, WeeklyReportsExportModal } from '../components';
+import { Header, HeaderAvatarTitle } from '../components';
+import { WeeklyReportsExportModal } from '../containers/Modals';
 import { CountryRoutes } from '../routes/CountryRoutes';
 import { countryFlagImage } from '../utils';
 import { checkIsMultiCountryUser } from '../store';
